pyproject.toml
[tool.poetry]
name = "acru-l"
version = "0.0.0"
description = ""
authors = ["Anthony Almarza <anthony.almarza@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "acru_l", from = "src" },
]
include = [
    "src/acru_l/**/requirements.txt",
]


[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "1.79.0"
"aws-cdk.custom-resources" = "1.79.0"
"aws-cdk.aws-secretsmanager" = "1.79.0"
"aws-cdk.aws-s3" = "1.79.0"
"aws-cdk.aws-cloudfront" = "1.79.0"
"aws-cdk.aws-lambda" = "1.79.0"
"aws-cdk.aws-apigatewayv2" = "1.79.0"
"aws-cdk.aws-route53" = "1.79.0"
"aws-cdk.aws-events" = "1.79.0"
"aws-cdk.aws-events-targets" = "1.79.0"
"aws-cdk.aws-dynamodb" = "1.79.0"
"aws-cdk.aws-rds" = "1.79.0"
environs = "^9.2.0"
pydantic = "^1.7.3"
"aws-cdk.aws-lambda-python" = "1.79.0"
"aws-cdk.aws-lambda-nodejs" = "1.79.0"
"aws-cdk.aws-apigatewayv2-integrations" = "1.79.0"
acru-l-toolkit = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6"
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.4.0"
pytest-responses = "^0.4.0"
pytest-picked = "^0.4.4"
pytest-instafail = "^0.4.2"
freezegun = "^1.0.0"
factory-boy = "^3.1.0"
codecov = "^2.1.11"
coverage = {extras = ["toml"], version = "^5.3"}
ipython = "^7.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = [
    "acru_l",
]
omit = [
    "tests/*",
    "factories/*"
]

[tool.coverage.report]
omit = [
    "tests/*",
    "factories/*"
]
show_missing = true
skip_empty = true
fail_under = 70

[tool.black]
line-length = 79

[tool.pytest.ini_options]
addopts = "-s"
python_files = [
    "test_*.py",
    "tests.py",
    "*_tests.py"
]
testpaths = [
    "./tests",
]
